okkk/requirements.txt
# requirements.txt
pytest
Appium-Python-Client
google-generativeai
langchain
langchain-google-genai
python-dotenv

